Changed bg_brightness type to float, removed redundant cast, added <precision> of 5.0 and improved tooltip
(bzr r3753.11.4)

+++ plugins/firepaint/src/firepaint.cpp
@@ -443,7 +443,7 @@ FireScreen::clear (CompAction         *action,
 void
 FireScreen::preparePaint (int time)
 {
-    float bg = (float) optionGetBgBrightness () / 100.0;
+    float bg = optionGetBgBrightness () / 100.0;
 
     if (init && !points.empty ())
     {
